+++ packages/kit/src/runtime/client/remote-functions/form.svelte.js
@@ -2,18 +2,11 @@
 /** @import { RemoteFormInput, RemoteForm, RemoteQueryOverride } from '@sveltejs/kit' */
 /** @import { RemoteFunctionResponse } from 'types' */
 /** @import { Query } from './query.svelte.js' */
-import { app_dir, base } from '__sveltekit/paths';
+import { app_dir, base } from '$app/paths/internal/client';
 import * as devalue from 'devalue';
 import { DEV } from 'esm-env';
 import { HttpError } from '@sveltejs/kit/internal';
-import {
-	app,
-	remote_responses,
-	started,
-	_goto,
-	set_nearest_error_page,
-	invalidateAll
-} from '../client.js';
+import { app, remote_responses, _goto, set_nearest_error_page, invalidateAll } from '../client.js';
 import { tick } from 'svelte';
 import { refresh_queries, release_overrides } from './shared.svelte.js';
 import { createAttachmentKey } from 'svelte/attachments';
@@ -42,7 +35,7 @@ export function form(id) {
 		let issues = $state.raw({});
 
 		/** @type {any} */
-		let result = $state.raw(started ? undefined : remote_responses[action_id]);
+		let result = $state.raw(remote_responses[action_id]);
 
 		/** @type {number} */
 		let pending_count = $state(0);
@@ -237,7 +230,7 @@ export function form(id) {
 
 				event.preventDefault();
 
-				const form_data = new FormData(form);
+				const form_data = new FormData(form, event.submitter);
 
 				if (DEV) {
 					validate_form_data(form_data, clone(form).enctype);
@@ -347,7 +340,7 @@ export function form(id) {
 				event.stopPropagation();
 				event.preventDefault();
 
-				const form_data = new FormData(form);
+				const form_data = new FormData(form, target);
 
 				if (DEV) {
 					const enctype = target.hasAttribute('formenctype')
@@ -355,10 +348,6 @@ export function form(id) {
 						: clone(form).enctype;
 
 					validate_form_data(form_data, enctype);
-				}
-
-				if (target.name) {
-					form_data.append(target.name, target?.getAttribute('value') ?? '');
 				}
 
 				await handle_submit(form, form_data, callback);
@@ -429,12 +418,12 @@ export function form(id) {
 			},
 			validate: {
 				/** @type {RemoteForm<any, any>['validate']} */
-				value: async ({ includeUntouched = false } = {}) => {
+				value: async ({ includeUntouched = false, submitter } = {}) => {
 					if (!element) return;
 
 					const id = ++validate_id;
 
-					const form_data = new FormData(element);
+					const form_data = new FormData(element, submitter);
 
 					/** @type {readonly StandardSchemaV1.Issue[]} */
 					let array = [];