+++ packages/package/test/fixtures/resolve-alias/src/lib/baz.ts
@@ -1,5 +1,0 @@
-export interface Baz {
-	baz: string;
-}
-
-export const baz: Baz = { baz: 'baz' };